feat: add ProductEventPublisher to publish product creation events and update quantity type to Integer

+++ src/main/java/com/phumlanidev/productservice/dto/ProductDto.java
@@ -23,7 +23,7 @@ public class ProductDto {
   @DecimalMin(value = "0.01", message = "Price must be greater than 0")
   private BigDecimal price;
   @DecimalMin(value = "1", message = "Quantity must be at least 1")
-  private BigDecimal quantity;
+  private Integer quantity;
   private String imageUrl;
 
 }
